+++ src/magnet.rs
@@ -120,9 +120,11 @@ impl std::error::Error for MagnetLinkError {
 ///
 /// More information is specified in [BEP-0009](https://bittorrent.org/beps/bep_0009.html), and
 /// even more appears in the wild, as explained [on Wikipedia](https://en.wikipedia.org/wiki/Magnet_URI_scheme).
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, Deserialize, Serialize)]
 #[cfg_attr(feature = "sea_orm", derive(sea_orm::DeriveValueType))]
 #[cfg_attr(feature = "sea_orm", sea_orm(value_type = "String"))]
+#[serde(try_from = "String")]
+#[serde(into = "String")]
 pub struct MagnetLink {
     /// Only mandatory field for magnet link parsing, unless the
     /// `magnet_force_name` crate feature is enabled.
@@ -337,6 +339,20 @@ impl FromStr for MagnetLink {
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         Self::new(s)
+    }
+}
+
+impl TryFrom<String> for MagnetLink {
+    type Error = MagnetLinkError;
+
+    fn try_from(s: String) -> Result<Self, Self::Error> {
+        Self::new(&s)
+    }
+}
+
+impl From<MagnetLink> for String {
+    fn from(m: MagnetLink) -> Self {
+        m.to_string()
     }
 }
 
@@ -578,5 +594,14 @@ mod tests {
             .map(|tracker| tracker.url().to_string())
             .collect::<Vec<_>>();
         assert!(found.is_empty());
+    }
+
+    #[test]
+    fn serialization_roundtrip() {
+        let magnet_url = std::fs::read_to_string("tests/bittorrent-v2-test.magnet").unwrap();
+        let magnet = MagnetLink::new(&magnet_url).unwrap();
+        let json_url = serde_json::to_string(&magnet_url).unwrap();
+        let deserialized_magnet: MagnetLink = serde_json::from_str(&json_url).unwrap();
+        assert_eq!(deserialized_magnet, magnet,);
     }
 }